requirements.txt
Openpyxl
streamlit
mysql-connector-python